Modules/FileDataWriter: make m_bytes_written unsigned
We cannot write a negative amount of bytes to file, after all.

+++ src/Modules/FileDataWriter/FileDataWriterModule.hpp
@@ -85,7 +85,7 @@ class FileDataWriterModule : public daqling::core::DAQProcess, public daqling::c
   std::atomic<bool> m_stopWriters;
 
   // Metrics
-  mutable std::atomic<int> m_bytes_written;
+  mutable std::atomic<size_t> m_bytes_written;
   mutable std::atomic<size_t> m_payload_queue_size;
   mutable std::atomic<size_t> m_payload_queue_bytes;
